services/embedding/requirements.txt
kafka-python==2.0.2
#openai==1.6.1
ollama==0.4.4
langchain==0.1.16
